+++ sktime/datasets/__init__.py
@@ -15,16 +15,16 @@ __all__ = [
     "load_PBS_dataset",
 ]
 
-from sktime.datasets.base import load_airline
-from sktime.datasets.base import load_gunpoint
-from sktime.datasets.base import load_arrow_head
-from sktime.datasets.base import load_basic_motions
-from sktime.datasets.base import load_osuleaf
-from sktime.datasets.base import load_italy_power_demand
-from sktime.datasets.base import load_japanese_vowels
-from sktime.datasets.base import load_longley
-from sktime.datasets.base import load_lynx
-from sktime.datasets.base import load_shampoo_sales
-from sktime.datasets.base import load_uschange
-from sktime.datasets.base import load_UCR_UEA_dataset
-from sktime.datasets.base import load_PBS_dataset
+from sktime.datasets.data_io import load_airline
+from sktime.datasets.data_io import load_gunpoint
+from sktime.datasets.data_io import load_arrow_head
+from sktime.datasets.data_io import load_basic_motions
+from sktime.datasets.data_io import load_osuleaf
+from sktime.datasets.data_io import load_italy_power_demand
+from sktime.datasets.data_io import load_japanese_vowels
+from sktime.datasets.data_io import load_longley
+from sktime.datasets.data_io import load_lynx
+from sktime.datasets.data_io import load_shampoo_sales
+from sktime.datasets.data_io import load_uschange
+from sktime.datasets.data_io import load_UCR_UEA_dataset
+from sktime.datasets.data_io import load_PBS_dataset
